prompt_skeleton_service/pyproject.toml
[project]
name = "prompt-skeleton-service"
version = "0.1.0"
description = "Config-driven FastAPI prompt skeleton service for Dify."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
